requirements.txt
jupyter==1.1.1
python-dotenv==1.0.1
httpx==0.27.2 
numpy==1.26.4
pandas==2.2.3
openpyxl==3.1.5
matplotlib==3.9.2
pypdf==5.1.0
pytest==8.3.3
openai==1.54.4
tiktoken==0.8.0
Flask==3.1.0
tavily-python==0.5.0
pydantic==2.9.2
llama-index==0.11.23
langchain-core==0.3.18
langchain-community==0.3.7
langchain-openai==0.2.8
langchain-text-splitters==0.3.2
langgraph==0.2.48
markdown==3.6
bleach==6.1.0
